Fixed build warnings caused by incompatible pointers types
This patch fixes build warnings caused by pointers being passed as
function parameters.  The compiler warns about pointer conversion
other than to void *, so e.g. VOID ** parameters need a cast.

This change is across a large number of files but does not
affect functionality.

The primary rationale for this fix is to clean up a mass of build
warnings so the useful warnings become apparent.

Change-Id: I9daffe5eae4c8fa9f129314223a460b3e701d4cf
Signed-off-by: Girish Pathak <girish.pathak@arm.com>