Fix assert error in SetInfo_Conf tests
This change fixes a problem in SetInfo_Conf test cases in which copying
of a new file name into FileInfo->FileName causes memory corruption.
The corruption is due to allocation of a small buffer following truncation of
file name by the filesystem. This fix adds checks to allocate a new buffer
if the size of the new file name to be copied is longer than a file name
by the filesystem.

The issue only occurs on filesystems where the maximum length of a file
name is less than the allocated space, for example ArmBootMonFs

Change-Id: I096ee3a7551d4d16a6371bc887d1cb48a28e5e41
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Girish Pathak <girish.pathak@arm.com>