Fix ExitBootService_Conf test case
A test case must write a recovery record before system reset
so that after reset the SCT framework finds the recovery record and
resumes the test case which was interrupted. If a test case does not
write any recovery record the SCT framework will simply flag the
case as failed and skip resumption of the test case.

This fix uses WriteResetRecord which creates the recovery record
and ReadResetRecord to read back and validate it.

Change-Id: I31baffa5da10376ccc89a2dec5d3f3b228d22618
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Girish Pathak <girish.pathak@arm.com>