Fix crash in Statistics_Func test case.
As per the UEFI specification if the size of the statistics table specified
by StatisticsSize is not big enough for all the statistics that are
collected by the network interface, then a partial buffer of statistics is
returned in StatisticsTable and StatisticsSize is set to the size required
to collect all the available statistics. Subsequent calls to Statistics will
result in invalid memory access and system crash if the previous value of
Statisticsize is larger than the size of EFI_NETWORK_STATISTICS.

This fix resets the value of StatisticsSize to the size of
EFI_NETWORK_STATISTICS to avoid memory corruption by Statistics call.

Change-Id: I7ad025afc338606cd6537b7f51e6efd44a20f5cd
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Girish Pathak <girish.pathak@arm.com>