SbbrBootServices: Bugfix for MemoryMap Check Test.
As per SBBR specification, "A UEFI runtime environment compliant with
SBBR must not be written with any assumption of an identity mapping
between virtual and physical memory maps."

Test case implementation was failing the test, if it is not identity
mapped, which is incorrect.

Corrected test case to warn the user that UEFI runtime environment is
identity mapped, instead of failure.

Signed-off-by: Supreeth Venkatesh <supreeth.venkatesh@arm.com>
Reported-by: Felix Poludov <Felixp@ami.com>