uefi-sct/SctPkg:Fix flaw in BBTestCreateEventEx_Func_Sub3
The intention of test is to validate the signal sequence among
three events with gEfiEventMemoryMapChangeGuid and different
Tpl. The call of AllocatePages() causes memorymap change and
trigger event Notify.
But the test has an assumption that CreateEventEx() will not
cause memorymap change itself, which is not true. When memory
pool is out of resource, the memory service will be called to
reserve more memory pool.
The fix is to filter the exception and only leave memorymap
change caused by AllocatePages() in test.

Cc: Supreeth Venkatesh <supreeth.venkatesh@arm.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Eric Jin <eric.jin@intel.com>
Reviewed-by: Supreeth Venkatesh <supreeth.venkatesh@arm.com>