uefi-sct/SctPkg:Implement iSCSI DevicePathToText Test
It is the new checkpoint of iSCSI DevicePathToText.
The Lun display is to follow with the UEFI Spec that
"The LUN is an 8 byte array that is displayed in
hexadecimal format with byte 0 first (i.e., on the
left) and byte 7 last (i.e, on the right).

Cc: Supreeth Venkatesh <supreeth.venkatesh@arm.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Eric Jin <eric.jin@intel.com>
Reviewed-by: Supreeth Venkatesh <supreeth.venkatesh@arm.com>