uefi-sct:SctPkg:Fix underflow in EraseBlocks test
The test case does not handle properly when
the 'EraseLengthGranularity' field of an Erase
Block Protocol instance is 1(SD device).
The Size parameter will underflow and become a
very large value, cause to exceed the last LBA.
The fix considers this case.
Besides, the redundant whitespaces on the tail
of line have been cleaned.

Cc: Supreeth Venkatesh <supreeth.venkatesh@arm.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Eric Jin <eric.jin@intel.com>
Reviewed-by: Supreeth Venkatesh <supreeth.venkatesh@arm.com>