uefi-sct/SctPkg: SNP Statistics may be unsupported
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=2160

If EFI_SIMPLE_NETWORK.Statistics() is not supported, it will return
EFI_UNSUPPORTED. In this case do not expect input parameters to be checked
for conformance.

If an error in SnpInterface->Start() occurs we error out. Therefore create
assertion messages always directly after each individual check.

Fix some typos in comments and messages.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>
Reviewed-by: Eric Jin <eric.jin@intel.com>