uefi-sct/SctPkg: setting key toggle state may be unsupported
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=2169

The UEFI specification allows EFI_SIMPLE_TEXT_INPUT_EX_PROTOCOL.SetState()
to return EFI_UNSUPPORTED if a state change is not supported. This for
instance may be the case when connecting via a serial connection.

Up to now the SCT has marked the conformance test as FAILED if SetState()
was called with invalid parameters and EFI_UNSUPPORTED was returned which
contradicts the specification.

Instead create a warning if SetState() returns EFI_UNSUPPORTED. This allows
the user to check if the non-support is justified.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>
Reviewed-by: Eric Jin <eric.jin@intel.com>