Check return status on calls to GetTime()
Not all platforms implement GetTime(), but the SCT just assumes calls to GetTime will be successful. If GetTime() doesn't return EFI_SUCCESS, then the EFI_TIME value will be uninitialized data.

Fix by checking the GetTime() return code. If it doesn't return EFI_SUCCESS, then use the traditional 1/1/1970 epoch so that the test report at least looks sane, but it is obvious that we don't have a valid timestamp.

Fixes: https://bugzilla.tianocore.org/show_bug.cgi?id=2870

Cc: G Edhaya Chandran <Edhaya.Chandran@arm.com>
Cc: Heinrich Schuchardt <xypron.glpk@gmx.de>
Cc: Samer El-Haj-Mahmoud <samer.el-haj-mahmoud@arm.com>
Signed-off-by: Grant Likely <grant.likely@arm.com>

Reviewed-by: Heinrich Schuchardt <xypron.glpk@gmx.de>
Reviewed-by: Samer El-Haj-Mahmoud <Samer.El-Haj-Mahmoud@arm.com>
Reviewed-by: G Edhaya Chandran <edhaya.chandran@arm.com>