uefi-sct/SctPkg: NULL deref in DevicePathToText test
Ref: https://bugzilla.tianocore.org/show_bug.cgi?id=3029

Function DevicePathToTextConvertDeviceNodeToTextCoverageTest() tests if
DeviceNodeToText() correctly converts a Relative Offset Range node. After calling SctConvertTextToDeviceNode() it tries to set the field Reserved of the returned device node to 0.

If the tested firmware does not return the expected text
SctConvertTextToDeviceNode() may return NULL or a device node that is shorter than expected. In both cases it is not possible to access the field Reserved.

So we must check both that the returned node is not NULL and that it has the exepected size.

Due to the missing check a NULL dereference was observed when running the SCT on U-Boot.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>
Reviewed-by: G Edhaya Chandran<edhaya.chandran@arm.com>
Reviewed-by: Grant Likely <grant.likely@arm.com>