edk2-test: Add support for building extra packages
The build.sh script is very useful for setting up the build environment
before calling the package build. Sometimes additional packages are
needed when building the SCT. (e.g., it is useful to build ShellPkg).
Refactor the build code to allow additional DSCs to be added to the
build.

This patch is useful when building a full SCT image that includes the
EDK2 shell.

Signed-off-by: Grant Likely <grant.likely@arm.com>
Cc: G Edhaya Chandran <Edhaya.Chandran@arm.com>
Cc: Barton Gao <gaojie@byosoft.com.cn>
Cc: Samer El-Haj-Mahmoud <samer.el-haj-mahmoud@arm.com>

Reviewed-by: G Edhaya Chandran<edhaya.chandran@arm.com>
Reviewed-by: Samer El-Haj-Mahmoud <Samer.El-Haj-Mahmoud@arm.com>