ImageIndex need be corrected to range 1 ~ DescriptorCount.
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=3272

The goal of the Support Routine: CheckForSupportGetImage is to detect the
current FMP instance's capability for supporting GetImage().

In current code, "ImageIndex = 0" is the input parameter, but it is
inconsistent with spec -"A unique number identifying the firmware image(s)
within the device. The number is between 1 and DescriptorCount."

It is a bug and need the fix.
The similar errors exist in CheckForSupportSetImage/CheckForSupportCheckImage.

Cc: G Edhaya Chandran <Edhaya.Chandran@arm.com>
Cc: Barton Gao <gaojie@byosoft.com.cn>
Cc: Samer El-Haj-Mahmoud <Samer.El-Haj-Mahmoud@arm.com>
Cc: Arvin Chen <arvinx.chen@intel.com>
Signed-off-by: Eric Jin <eric.jin@intel.com>

Reviewed-by: Barton Gao <gaojie@byosoft.com.cn>
Reviewed-by: G Edhaya Chandran<edhaya.chandran@arm.com>