uefi-sct/SctPkg: Update page alignment calculations
This is to fix the SCT BS.AllocatePages failures (not found) with the case that the Start address is not aligned to 64k.
For example,
  The following is available memory region for testing:
    0000000082012000-00000000EB6D9FFF 00000000000696C8
  With the current page alignment calculation, we will get:
    Start address is 0x82020000
    PageNum is 0x696B8
  In BS.AllocatePages, it will make the end address align with 64k,
  so PageNum will be changed from 0x696B8 to 0x696C0. Therefore, the
  end address will become 0xEB6E0000 which is larger than 0xEB6D9FFF,
  so we get not found error in the end.

Therefore, the calculation for getting the PageNum should be updated to PageNum - (2 * EFI_SIZE_TO_PAGES(0x10000)) so that we won't get a wrong PageNum to allocate a memory with a size larger than available space's size.

With this solution, the example above will get 0x696A8 as calculated PageNum. Then, in BS.AllocatePages, the PageNum will be changed from
0x696A8 to 0x696B0. Therefore, the end address will become 0xEB6D0000 that is smaller than 0xEB6D9FFF, so we get not found error in the end.

I also tested this solution on two ARM platforms (NXP1046A and RPi4).

Cc: Samer El-Haj-Mahmoud <samer.el-haj-mahmoud@arm.com>
Cc: G Edhaya Chandran <edhaya.chandran@arm.com>
Cc: Barton Gao <gaojie@byosoft.com.cn>
Signed-off-by: Sunny Wang <sunny.wang@arm.com>

Reviewed-by: G Edhaya Chandran<edhaya.chandran@arm.com>
Reviewed-by: Tuan Phan <tuanphan@os.amperecomputing.com>