uefi-sct/SctPkg: invalid values for SetWakeupTime()
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=2529

According to the UEFI specification EFI_TIME.YEAR may take values in the
range 1900-9999. The specification does not forbid to set a wakeup time in
the past. We should not expect EFI_INVALID_PARAMETER to be returned by
SetWakeupTime() for the years 1997 and 2100.

Values 1899 and 10000 must lead to this error code.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>
Reviewed-by: G Edhaya Chandran<edhaya.chandran@arm.com>