Update capsule header guid
UpdateCapsule() Conformance requires that a capsule marked
CAPSULE_FLAGS_PERSIST_ACROSS_RESET must also provide a valid
ScatterGatherList. In some platforms, however, if a capsule is marked as
such, the capsule is validated before being persisted in memory (which
is when the ScatterGatherList is verified). Due to this, those platforms
fail the SCT that validates this conformance as they return
EFI_UNSUPPORTED as the capsule header guid being passed to the function
is not a known valid capsule header guid.

This commit updates the capsule header guid used in the
MiscRuntimeService tests to use a known good capsule header guid,
gWindowsUxCapsuleGuid, which causes these platforms to pass this SCT.