MiscBootServices: Stall_Func: Reduces the stall interval for Stall_Func
The Stall_Func test on the highest TPL causes issues with the disk IO by blocking interrupts. This blocking can cause disk corruption through IO timeouts. Since this doesn't seem to be the intent of the test this revision reduces the stall interval from 10 seconds to 4 and adjusts the delta tolerance in scale.

Signed-off-by: Robert Wood <rwood.ce@comcast.net>

Cc: Samer El-Haj-Mahmoud <Samer.El-Haj-Mahmoud@arm.com>

Reviewed-by: G Edhaya Chandran <edhaya.chandran@arm.com>
Reviewed-by: Samer El-Haj-Mahmoud <Samer.El-Haj-Mahmoud@arm.com>