uefi-sct/SctPkg: update path to edk2-test-parser, check for repo
buildzip.sh was referencing the obsolete sct_parser project
which has been replaced with edk2-test-parser.

The script now checks for the edk2-test-parser repo in $pwd,
if not present then the script exits and reminds the user to
clone edk2-test-parser.
Now the script copies edk2-test-parser/* recursively to the repo.

Cc: G Edhaya Chandran <Edhaya.Chandran@arm.com>
Cc: Barton Gao <gaojie@byosoft.com.cn>
Cc: Carolyn Gjertsen <Carolyn.Gjertsen@amd.com>

Signed-off-by: Abhi Singh <Abhi.Singh@arm.com>