UEFI-SCT Case spec conversion to markdown
Previously UEFI-SCT case spec was maintained as a word document source and
released as PDF. In order for the document to be user-friendly and easy for
development and maintenence, the source has been converted to markdown.
Conversion done by Prachotan Reddy Bathi <prachotan.bathi@arm.com>
The old 'Src' directory has been deleted.

As part of this commit the old .docx and .pdf have been deleted
Facilities to support readthedocs in the future has been added.

Co-authored-by: Prachotan Reddy Bathi <prachotan.bathi@arm.com>
Reviewed-by: G Edhaya Chandran <edhaya.chandran@arm.com>