SctPkg: Drop ArmCompilerIntrinsicsLib
edk2 PR tianocore/edk2#6048 moves
ArmCompilerIntrinsicsLib to MdePkg and into MdeLibs.dsc.inc.

This patch drops all of the references to the ArmPkg location as
these *_SCT.dsc are now getting it from MdeLibs.dsc.inc. This is
done in one shot as it is a breaking change and has no functional
impact.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>